Upgraded idock from 1.5 to 1.6

+++ idock/common.hpp
@@ -36,7 +36,7 @@ namespace idock
 	using boost::filesystem::path;
 
 	/// Use single precision floating point due to the fact that only 4GB RAM is installed on Mac workstations.
-	typedef float fl;
+	typedef double fl;
 
 	// Choose the appropriate Mersenne Twister engine for random number generation on 32-bit or 64-bit platform.
 #if defined(__x86_64) || defined(__x86_64__) || defined(__amd64) || defined(__amd64__) || defined(_M_X64) || defined(_M_AMD64)
